requirements.txt
telethon>=1.28.5
python-dotenv>=1.0.0
openpyxl>=3.1.0
aiofiles>=23.1.0
cryptography>=41.0.0
matplotlib>=3.7.0
pandas>=2.0.0
seaborn>=0.12.0

# Web Panel Dependencies
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-login>=0.6.0
flask-wtf>=1.1.0
wtforms>=3.0.0
flask-limiter>=3.3.0
flask-jwt-extended>=4.5.0
werkzeug>=2.3.0